lb_factory::contract
Module execute

Functions

_set_lb_hooks_parameters_on_pair

Internal function to set a hooks contract to the pair

add_quote_asset

Function to add an asset to the whitelist of quote assets

create_lb_pair

Creates a liquidity bin LbPair for token_x and token_y.

create_lb_pair_part2

Handles the reply from instantiating the new LbPair.

force_decay

Function to force the decay of the volatility accumulator of a pair. Needs to be called by the owner.

remove_lb_hooks_on_pair

Function to remove the hooks contract from the pair. Needs to be called by an address with the LB_HOOKS_MANAGER_ROLE. Reverts if:

remove_preset

Remove the preset linked to a bin step. Needs to be called by the owner.

remove_quote_asset

Function to remove an asset from the whitelist of quote assets

set_fee_parameters_on_pair

Function to set the fee parameters of a LbPair. Needs to be called by the owner.

set_fee_recipient

Function to set the recipient of the fees. This address needs to be able to receive SNIP20s.

set_flash_loan_fee

Function to set the flash loan fee. Needs to be called by the owner. Reverts if:

set_lb_hooks_parameters_on_pair

Function to set the hooks parameters of a pair. Needs to be called by an address with the LB_HOOKS_MANAGER_ROLE. Reverts if:

set_lb_pair_ignored

Function to set whether the pair is ignored or not for routing, it will make the pair unusable by the router. Needs to be called by the owner.

set_lb_pair_implementation

Set the LbPair implementation details.

set_lb_token_implementation

Set the LbToken implementation details.

set_preset

Sets the preset parameters of a bin step Needs to be called by the owner.

set_preset_open_state

Sets if the preset is open or not to be used by users. Needs to be called by the owner.