lb_factory::contract::execute
Function set_preset

pub fn set_preset(
    deps: DepsMut<'_>,
    _env: Env,
    info: MessageInfo,
    bin_step: u16,
    base_factor: u16,
    filter_period: u16,
    decay_period: u16,
    reduction_factor: u16,
    variable_fee_control: u32,
    protocol_share: u16,
    max_volatility_accumulator: u32,
    is_open: bool,
) -> Result<Response>

Sets the preset parameters of a bin step Needs to be called by the owner.

Reverts if:

The binStep is lower than the minimum bin step

Arguments

bin_step - The bin step in basis point, used to calculate the price

base_factor - The base factor, used to calculate the base fee, baseFee = baseFactor * binStep

filter_period - The period where the accumulator value is untouched, prevent spam

decay_period - The period where the accumulator value is decayed, by the reduction factor

reduction_factor - The reduction factor, used to calculate the reduction of the accumulator

variable_fee_control - The variable fee control, used to control the variable fee, can be 0 to disable it

protocol_share - The share of the fees received by the protocol

max_volatility_accumulator - The max value of the volatility accumulator

is_open - Whether the preset is open or not to be used by users