lb_factory::contract
Module query

Functions

get_all_bin_steps

Returns the list of available bin steps with a preset.

get_all_lb_pairs

Returns all the LbPair of a pair of tokens.

get_fee_recipient

Get the protocol fee recipient.

get_flash_loan_fee

Get the fee for flash loans, in 1e18.

get_lb_pair_at_index

Returns the LbPair created at the given index.

get_lb_pair_implementation

Get the code ID and hash of the LbPair implementation.

get_lb_pair_information

Returns the LbPairInformation if it exists, if not, then empty information is returned. The order doesn’t matter.

get_lb_token_implementation

Get the code ID and hash of the LbToken implementation.

get_max_flash_loan_fee

Get the maximum fee percentage for flashLoans.

get_min_bin_step

Get the minimum bin step a pair can have.

get_number_of_lb_pairs

Returns the number of LbPairs created.

get_number_of_quote_assets

Returns the number of quote assets whitelisted.

get_open_bin_steps

Returns the list of open bin steps.

get_preset

Returns the different parameters of the preset.

get_quote_asset_at_index

Returns the quote asset whitelisted at the given index.

is_quote_asset

Returns whether a token is a quote asset (true) or not (false).