lb_factory::contract
Module state

Structs

EphemeralLbPair

State

Enums

ContractStatus

Constants

EPHEMERAL_LB_PAIR

Statics

ALL_LB_PAIRS

AVAILABLE_LB_PAIR_BIN_STEPS

Mapping from a (tokenA, tokenB) to a set of available bin steps, this is used to keep track of the bin steps that are already used for a pair. The tokens are ordered to save gas, but they can be in the reverse order in the actual pair. Always query one of the 2 tokens of the pair to assert the order of the 2 tokens.

CONTRACT_STATUS

FEE_RECIPIENT

FLASH_LOAN_FEE

LB_PAIRS_INFO

Mapping from a (tokenA, tokenB, binStep) to a LBPair. The tokens are ordered to save gas, but they can be in the reverse order in the actual pair. Always query one of the 2 tokens of the pair to assert the order of the 2 tokens.

LB_PAIR_IMPLEMENTATION

LB_TOKEN_IMPLEMENTATION

PRESETS

PRESET_BIN_STEPS

QUOTE_ASSET_WHITELIST

STATE